Add files via upload
These are the PDF files we initially used to keep track of data-port statuses and equipment information and location. They're inaccessible, cluttered with irrelevant information, and difficult to maintain. The only reason I'm comfortable putting them online is because even though the data-port and equipment info was updated in 2020, the physical layout of the library is completely different, so most of the map is useless to anyone.