Update DeviceNetworkDatabase.sql
Made some pretty big changes to the structure of 'graveyard'; adding pricing for staff/office devices that have been withdrawn and adding the 'Recallable' field to differentiate between devices that are gone for good and those that are temporarily stored. Changed all the data stored as bits to tinyint(1) because it had caused some problems when recreating the database in other environments. Also added 'fake_staff' and 'fake_departments' so I could demonstrate functionality for managing office equipment without exposing my colleagues' emails and such.  So, just for funsies, this library is populated solely by me and characters from the manga/anime series "Jojo's Bizarre Adventure.'