Update placeholder.php
Changed the way PHP creates $group_number to pull directly from the database instead of doing a ++ loop. Should have been obvious to do it that way from the start, mea culpa. Added  the branch's location codes to the sort in order to support the creation of JSON for the branch library as well.